feat: Move external api tests to the itests
